+++ plugins/cjs/ui/shared/config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  for: function(routeName) {
-    return window.tinydoc.getRuntimeConfigs('cjs').filter(function(config) {
-      return config.routeName === routeName;
-    })[0];
-  }
-};