+++ plugins/markdown/ui/shared/constants.js
@@ -1,1 +1,0 @@
-exports.CFG_CLASS_BROWSER_GROUP_BY_FOLDER = 'markdown:classBrowser:groupByFolder';